Flesh out barebones example
This is with very little library support from a third-party flake which
will surely exist.

The idea is to first do it all the laborous copy+paste heavy way, and
then abstract a nicer interface out later.

The real main motivation of this is just to make everything else work at
all.